requirements.txt
colorama==0.4.6
pygame==2.5.2
python-telegram-bot==20.7
pyngrok==7.2.3
flask
flask-cors
gunicorn
cryptography
requests
Pillow==10.1.0
